fix(cards): renames cards to make more sense
